requirements.txt
langchain
langgraph
langchain_groq
langchain-chroma
jq
langchain-community 
faiss-cpu
langchain_ollama
langchain_openai